+++ src/js/components/forms/fancy/single-value.js
@@ -1,25 +1,20 @@
-import React from 'react';
+import React from 'React';
+var classes = require('classnames');
 
-const Type = React.PropTypes;
-
-var SingleValue = React.createClass({
-	displayName: 'SingleValue',
-
-	propTypes: {
-		placeholder: React.PropTypes.string, // this is default value provided by React-Select based component
-		value: React.PropTypes.object // selected option
+export default React.createClass({
+  	propTypes: {
+		placeholder: React.PropTypes.string,       // this is default value provided by React-Select based component
+		value: React.PropTypes.object              // selected option
 	},
-	render: function render() {
+	render: function() {
 
 		var classNames = classes('Select-placeholder', this.props.value && this.props.value.className);
-		return React.createElement(
-			'div',
-			{
-				className: classNames,
-				style: this.props.value && this.props.value.style,
-				title: this.props.value && this.props.value.title
-			},
-			this.props.placeholder
+		return (
+			<div
+				className={classNames}
+				style={this.props.value && this.props.value.style}
+				title={this.props.value && this.props.value.title}
+				>{this.props.placeholder}</div>
 		);
 	}
 });
